Sat, Oct 19, 2024, 11:48 AM -03:00

+++ 0/concatInt.c
@@ -10,7 +10,7 @@ int* plusOne(int* digits, int digitsSize, int* returnSize) {
     }
     num = num + 1;
 
-    int tempNum = num;
+    long int tempNum = num;
     int digitCount = 0;
     while (tempNum > 0) {
         tempNum /= 10;
